1. fix all actions and tests 2. fix js and jsp about getProject from session. 3. remove 'Move Story' from 'Release Plan' page. 4. remove useless actions 5. add new http libraries.

+++ java/ntut/csie/ezScrum/web/helper/ProductBacklogHelper.java
@@ -93,11 +93,16 @@ public class ProductBacklogHelper {
 	 * @return StringBuilder
 	 */
 	public StringBuilder deleteStory(long storyId) {
-		removeTask(storyId);
-		mProductBacklogMapper.deleteStory(storyId);
-
+		StoryObject story = getStory(storyId);
 		StringBuilder result = new StringBuilder("");
-		result.append("{\"success\":true, \"Total\":1, \"Stories\":[{\"Id\":" + storyId + "}]}");
+
+		if (story != null) {
+			removeTask(storyId);
+			mProductBacklogMapper.deleteStory(storyId);
+			result.append("{\"success\":true, \"Total\":1, \"Stories\":[{\"Id\":" + storyId + "}]}");
+		} else {
+			result.append("{\"success\":false, \"Total\":0, \"Stories\":[{}]}");
+		}
 		return result;
 	}
